Add support for storing embeddings in ARRAY(REAL) columns
(cherry picked from commit da0e3c3b584fab34b69180b6e152f3b3e41d90ec)

+++ plugin/starburst-ai/src/main/java/com/starburstdata/trino/plugin/ai/EmbeddingModelClient.java
@@ -21,5 +21,5 @@ public interface EmbeddingModelClient
 {
     List<Double> generateEmbedding(Slice sourceString);
 
-    List<List<Double>> generateEmbeddings(List<Slice> sourceStrings);
+    List<List<Float>> generateEmbeddings(List<Slice> sourceStrings);
 }
